Module Polars

module Common : sig ... end

module Data_frame : sig ... end

module Data_type : sig ... end

module Env : sig ... end

There are many environment variables that effect Polars behavior, mostly related to the display of tables. This module is responsible for setting the default values of these variables so that e.g. expect tests will be less fragile.

module Expr : sig ... end

An Expr.t is the basic building block of how to select columns from a dataframe. It is a representation of a lazy computation over a dataframe which can be executed via functions such as Lazy_frame.select/ Data_frame.select and Lazy_frame.with_columns/Data_frame.with_columns.

module Fill_null_strategy : sig ... end

module Lazy_frame : sig ... end

module Schema : sig ... end

module Series : sig ... end

module Sql_context : sig ... end